move spaninfo code to walk.go
